Use `remq` instead of `delq` in `selectrum-prescient-create-and-bind-toggle-command`.
Even though we use `make-local-variable` to ensure that
`prescient-filter-method` is buffer-local in the minibuffer, it seems
that `delq` might still modify `prescient-filter-method` as a global
variable.  It was assumed that destructive functions would only modify
the buffer-local value, but that is apparently not the case.

See also issue #123 and PR #124.