Remove `prescient-completion-enable-sort` to avoid confusion.
It is better to make the various completion UIs use
`prescient-completion-sort` directly, as this allows the sorting of
candidates filtered by other completion styles and avoids Company
Prescient re-sorting candidates from `company-capf`.

To keep things simple, the option of modifying completion-table
metadata is removed.

See #129 and conversations in minad/vertico#237.